Update setParam()
Updates include using only table identifier, field number, and call to calculateOffset(). If there is a parameterized field for any column in the table, then setParam() is written for all the columns in the table to reduce the amount of generated code for additional INSERT calls to the same table. In the future, it should be updated to include more than int and string.